Add versions list CatalogClient tests.
Change functional CatalogClient tests
to use mock server.
Add helper functionality for mock server.

Relates-To: OLPEDGE-1606
